refactor(#28): authenticate 메소드 service -> repository로 분리
